+++ modules/api/gallery.py
@@ -184,7 +184,7 @@ def register_api(app: FastAPI): # register api
             folder = unquote(folder).replace('%3A', ':')
             t0 = time.time()
             numFiles = 0
-            files = files_cache.directory_files(folder, recursive=True)
+            files = files_cache.list_files(folder, recursive=True)
             # files = list(files_cache.directory_files(folder, recursive=True))
             # files.sort(key=os.path.getmtime)
             for f in files: